requirements.txt
flask
python-dotenv
pydantic
google-generativeai
werkzeug
